counter action buttons component
