+++ apps/thundermoon_web/assets/css/counter.scss
@@ -4,9 +4,6 @@
   grid-template-columns: repeat(2, 1fr);
 }
 
-.counter-controls {
-}
-
 .counter {
   display: grid;
   grid-gap: 0.5rem 0.5rem;
