+++ apps/thundermoon_web/lib/thundermoon_web/live/counter_live/digit.ex
@@ -3,11 +3,13 @@ defmodule ThundermoonWeb.CounterLive.Digit do
 
   alias Thundermoon.Counter
 
+  @impl true
   def handle_event("inc", _, socket) do
     Counter.inc(socket.assigns.digit)
     {:noreply, socket}
   end
 
+  @impl true
   def handle_event("dec", _, socket) do
     Counter.dec(socket.assigns.digit)
     {:noreply, socket}
